chatbot/requirements.txt
streamlit>=1.32.0
sentence-transformers>=2.7.0
chromadb>=0.4.22
pandas>=2.0.0
nltk>=3.8.1
requests>=2.31.0
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6
pydantic[email]>=2.5.0
pytest>=7.4.0
httpx>=0.25.0
python-dotenv>=1.0.0
PyJWT>=2.8.0
